Move exceptions into their own package and add checked exception for when tokens are invalid
