+++ Client/src/main/java/org/gusdb/oauth2/exception/ConflictException.java
@@ -1,4 +1,4 @@
-package org.gusdb.oauth2.client;
+package org.gusdb.oauth2.exception;
 
 public class ConflictException extends Exception {
