+++ Client/src/main/java/org/gusdb/oauth2/exception/InvalidTokenException.java
@@ -1,0 +1,9 @@
+package org.gusdb.oauth2.exception;
+
+public class InvalidTokenException extends Exception {
+
+  public InvalidTokenException(Throwable cause) {
+    super(cause);
+  }
+
+}
