+++ Server/src/main/java/org/gusdb/oauth2/config/ApplicationConfig.java
@@ -20,7 +20,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.gusdb.oauth2.InitializationException;
 import org.gusdb.oauth2.assets.StaticResource;
-import org.gusdb.oauth2.shared.CryptoException;
+import org.gusdb.oauth2.exception.CryptoException;
 import org.gusdb.oauth2.shared.SigningKeyStore;
 
 /** Example config JSON **
