+++ Server/src/main/java/org/gusdb/oauth2/service/OAuthService.java
@@ -60,12 +60,12 @@ import org.gusdb.oauth2.Authenticator.DataScope;
 import org.gusdb.oauth2.Authenticator.RequestingUser;
 import org.gusdb.oauth2.UserInfo;
 import org.gusdb.oauth2.assets.StaticResource;
-import org.gusdb.oauth2.client.ConflictException;
 import org.gusdb.oauth2.client.Endpoints;
 import org.gusdb.oauth2.client.HttpStatus;
-import org.gusdb.oauth2.client.InvalidPropertiesException;
 import org.gusdb.oauth2.client.OAuthClient;
 import org.gusdb.oauth2.config.ApplicationConfig;
+import org.gusdb.oauth2.exception.ConflictException;
+import org.gusdb.oauth2.exception.InvalidPropertiesException;
 import org.gusdb.oauth2.server.OAuthServlet;
 import org.gusdb.oauth2.service.util.AuthzRequest;
 import org.gusdb.oauth2.service.util.JerseyHttpRequestWrapper;
